+++ BusinessMudanzas/ICalculateMovements.cs
@@ -7,7 +7,6 @@ namespace BusinessMudanzas
     public interface ICalculateMovements
     {
         IList<WorkDay> GetLists(string fileString);
-        int GetTravels(WorkDay workday);
         int GetTravelsRecursive(WorkDay workday);
         string GetTotaltravels(IList<WorkDay> lists);
     }
