+++ WebApiMudanzas/lazy_loading_example_input.txt
@@ -1,0 +1,40 @@
+5
+4
+30
+30
+1
+1
+3
+20
+20
+20
+11
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+6
+9
+19
+29
+39
+49
+59
+10
+32
+56
+76
+8
+44
+60
+47
+85
+71
+91
